Pin versions to enable repeatable build
It was not possible to rebuild a working version of the app due to
issues with breaking changes in transient dependencies. This change pins
the versions to match the build currently deployed to production. There
are a number of upgrades required, but these will need to be completed
separately. The versions are pinned by using fixed versions in the
package.json, introducing a package-lock.json generated from the live
docker image and pinning the node version for the docker image.

In addition to fixing the build, this PR also:

* Adds an .nvmrc to specify the node version for local development
* Adds a Makefile with our standard targets to make this repo more
  accessible
* Adds a LICENCE file
* Updates the README inline with the changes and to fix the invalid POST
  command for testing whether the API is functional